verify_setup.py
import asyncio
from pathlib import Path
from dotenv import load_dotenv
import pandas as pd

class GoogleSheetsService:
    def __init__(self):
        # Create a new instance of Google Sheets Service
        pass

class VerifySetupService:
    async def verify_setup(self):
        # To be implemented. Currently empty.
        pass

if __name__ == '__main__':
    # Example usage of the VerifySetupService
    service = VerifySetupService()
    await service.verify_setup()